// frontend/src/types/ListRecipe.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface ListRecipe {
  ingredient_count: number
  id: number
  name: string
  description: string | null
  image_url: string | null
  servings: number | null
  work_time: number | null
  wait_time: number | null
  author: string
  created_at: number
}